Edit parameter description attribute to include age-specific definitions (helper function for `expand_params_age()`)

Edit parameter description attribute to include age-specific definitions (helper function for `expand_params_age()`)

expand_params_desc_age(params_desc)

Arguments

params_desc

parameter descriptions, as initialized in `read_params()`

Examples

params <- read_params("PHAC_testify.csv")
params_desc <- attr(params, "description")
expand_params_desc_age(params_desc)
#>                                                                                                                                                                                                                            beta0 
#>                                                                                     "Baseline (non-intervention) transmission across categories per age category (if single value, this same value is taken for all age groups)" 
#>                                                                                                                                                                                                                               Ca 
#>                                                                                                                                                                                "relative asymptomatic transmission (or contact)" 
#>                                                                                                                                                                                                                               Cp 
#>                                                                                                                                                                              "relative presymptomatic transmission (or contact)" 
#>                                                                                                                                                                                                                               Cm 
#>                                                                                                                                                                          "relative mildly symptomatic transmission (or contact)" 
#>                                                                                                                                                                                                                               Cs 
#>                                                                                                                                                                        "relative severely symptomatic transmission (or contact)" 
#>                                                                                                                                                                                                                            alpha 
#>                                                                                                                                                                                                 "Fraction of cases asymptomatic" 
#>                                                                                                                                                                                                                            sigma 
#>                                                                                                                                                                                                        "1/time in exposed class" 
#>                                                                                                                                                                                                                          gamma_a 
#>                                                                                                                                                                                               "1/time for asymptomatic recovery" 
#>                                                                                                                                                                                                                          gamma_m 
#>                                                                                                                                                                                         "1/time for mildly symptomatic recovery" 
#>                                                                                                                                                                                                                          gamma_s 
#>                                                                                                                                                                   "1/time for severely symptomatic transition to hospital/death" 
#>                                                                                                                                                                                                                          gamma_p 
#>                                                                                                                                                                                                "1/time in pre-symptomatic class" 
#>                                                                                                                                                                                                                              rho 
#>                                                                                                                                                                                                "1/time in hospital (acute care)" 
#>                                                                                                                                                                                                                            delta 
#>                                                                                                                                                                                    "Fraction of acute-care cases that are fatal" 
#>                                                                                                                                                                                                                               mu 
#>                                                                                                                                                                                    "Fraction of symptomatic cases that are mild" 
#>                                                                                                                                                                                                                                N 
#>                                                                                                                                                                                               "Population size per age category" 
#>                                                                                                                                                                                                                               E0 
#>                                                                                                                                                                                                         "Initial number exposed" 
#>                                                                                                                                                                                                                     nonhosp_mort 
#>                                                                                                                                                                               "probability of mortality without hospitalization" 
#>                                                                                                                                                                                                                            iso_m 
#>                                                                                                                                                                               "Relative self-isolation/distancing of mild cases" 
#>                                                                                                                                                                                                                            iso_s 
#>                                                                                                                                                                             "Relative self-isolation/distancing of severe cases" 
#>                                                                                                                                                                                                                             phi1 
#>                                                                                                                                                                                              "Fraction of hospital cases to ICU" 
#>                                                                                                                                                                                                                             phi2 
#>                                                                                                                                                                                                    "Fraction of ICU cases dying" 
#>                                                                                                                                                                                                                             psi1 
#>                                                                                                                                                                                                 "Rate of ICU back to acute care" 
#>                                                                                                                                                                                                                             psi2 
#>                                                                                                                                                                                                           "Rate of ICU to death" 
#>                                                                                                                                                                                                                             psi3 
#>                                                                                                                                                                                                  "Rate of post-ICU to discharge" 
#>                                                                                                                                                                                                                           c_prop 
#>                                                                                                                                                                               "fraction of incidence reported as positive tests" 
#>                                                                                                                                                                                                                     c_delay_mean 
#>                                                                                                                                                                                "average delay between incidence and test report" 
#>                                                                                                                                                                                                                       c_delay_cv 
#>                                                                                                                                                                                      "coefficient of variation of testing delay" 
#>                                                                                                                                                                                                                        proc_disp 
#>                                                                                                                                                                    "dispersion parameter for process error (0=demog stoch only)" 
#>                                                                                                                                                                                                                             zeta 
#>                                                                                                                                                                                       "phenomenological heterogeneity parameter" 
#>                                                                                                                                                                                                                          W_asymp 
#>                                                                                                                                                                                                          " asymptomatic weights" 
#>                                                                                                                                                                                                                              P_S 
#>                                                                                                                                                                                                  " probability of positive in S" 
#>                                                                                                                                                                                                                              P_E 
#>                                                                                                                                                                                                  " probability of positive in E" 
#>                                                                                                                                                                                                                             P_Ia 
#>                                                                                                                                                                                                  "probability of positive in Ia" 
#>                                                                                                                                                                                                                             P_Ip 
#>                                                                                                                                                                                                  "probability of positive in Ip" 
#>                                                                                                                                                                                                                             P_Im 
#>                                                                                                                                                                                                  "probability of positive in Im" 
#>                                                                                                                                                                                                                             P_Is 
#>                                                                                                                                                                                                  "probability of positive in Is" 
#>                                                                                                                                                                                                                              P_H 
#>                                                                                                                                                                                                  " probability of positive in H" 
#>                                                                                                                                                                                                                             P_H2 
#>                                                                                                                                                                                                 " probability of positive in H2" 
#>                                                                                                                                                                                                                           P_ICUs 
#>                                                                                                                                                                                               " probability of positive in ICUs" 
#>                                                                                                                                                                                                                           P_ICUd 
#>                                                                                                                                                                                                "probability of positive in ICUd" 
#>                                                                                                                                                                                                                              P_R 
#>                                                                                                                                                                                                   "probability of positive in R" 
#>                                                                                                                                                                                                                            omega 
#>                                                                                                                                                                                        "1/waiting time for tests to be returned" 
#>                                                                                                                                                                                                                testing_intensity 
#>                                                                                                                                                                                                       "tests per capita per day" 
#>                                                                                                                                                                                                                            iso_t 
#>                                                                                                                                                                         "Relative self-isolation/distancing of tested-positives" 
#>                                                                                                                                                                                                                      testing_tau 
#>                                                                                                                                                                                   "tests per capita per infected person per day" 
#>                                                                                                                                                                                                                             pmat 
#> "Contact probability matrix: the entry in row i and column j represents the probability that an individual of age i contacts an individual of age j (but not necessarily the other way around---only matrix rows must sum to 1)"