Simulation History

Simulation History

simulation_history(
  model,
  add_dates = TRUE,
  sim_params = NULL,
  include_initial_date = TRUE,
  obs_error = FALSE,
  condense = FALSE
)

Arguments

model

a flexmodel object

add_dates

should a column called Date be added?

sim_params

vector to pass to a TMB AD function -- you can get examples of these out of the opt_par element of a fitted flexmodel or by calling tmb_params_trans

include_initial_date

should the first row be the initial state vector, or the first date after the initial?

obs_error

should random observation error be added if it is defined in the model? -- see update_error_dist

condense

should a condensed set of simulation variables be returned? -- see update_condense_map

See also