Update Mistry-based parameters

Recalculate beta0 and/or pmat if setting-specific contact rates and/or overall transmissiblity are changed

update_params_mistry(
  params,
  contact_rate_setting = NULL,
  transmissibility = NULL
)

Arguments

params

parameter list initialized using `expand_params_mistry()`

contact_rate_setting

named list containing setting-specific contact rates (in units of average contacts in the given setting per individual of age i with individuals of age j per day)

transmissibility

probability of transmission upon contact with an infected (beta0 = transmissibility * contact_rate)

Value

an object of class `params_pansim`

Details

For `contact_rate_setting`, a full list (initialized with `mk_contact_rate_setting()`) or a partial list (e.g. `list(community = 0)`) can be provided. If a partial list is provided, default values from Mistry et al. 2021 are assumed for the contact rates that have not been specified (see `mk_contact_rate_setting()` for values).

Examples

params <- read_params("PHAC_testify.csv")
age_cat <- mk_agecats(min = 0, max = 80, da = 10)
params_mistry <- expand_params_mistry(params = params, province = "ON", age_cat = age_cat)
params_mistry <- update_params_mistry(params_mistry,
 contact_rate_setting = list(school = 0))